fix(release): restore v0.4.1 qualification
