+++ loopx/help_surface.py
@@ -201,8 +201,16 @@ COMMAND_GROUPS: list[dict[str, object]] = [
                 "purpose": "Resolve a portable weekly profile, evaluate report triggers, and compose provider-neutral run receipts.",
             },
             {"command": "loopx auto-research", "purpose": "Project public-safe research frontiers."},
+            {
+                "command": "loopx agent-turn-recall",
+                "purpose": "Prepare goal-scoped memory guidance for one admitted autonomous agent turn.",
+            },
             {"command": "loopx multi-agent", "purpose": "Launch visible role-scoped Codex TUI agents."},
             {"command": "loopx canary", "purpose": "Plan or run catalog-informed smoke profiles."},
+            {
+                "command": "loopx promotion-readiness",
+                "purpose": "Record release-scoped canary readiness evidence in the shared runtime ledger.",
+            },
             {"command": "loopx benchmark", "purpose": "Use fixture-only benchmark runner skeletons by default."},
         ],
     },
