+++ man/loopx.1
@@ -1,4 +1,4 @@
-.TH LOOPX 1 "" "LoopX 0.4.0" "User Commands"
+.TH LOOPX 1 "" "LoopX 0.4.1" "User Commands"
 .SH NAME
 loopx \- control-plane helper for long-running agent work
 .SH SYNOPSIS
@@ -186,11 +186,17 @@ Resolve a portable weekly profile, evaluate report triggers, and compose provide
 \fBloopx auto\-research\fR
 Project public\-safe research frontiers.
 .TP
+\fBloopx agent\-turn\-recall\fR
+Prepare goal\-scoped memory guidance for one admitted autonomous agent turn.
+.TP
 \fBloopx multi\-agent\fR
 Launch visible role\-scoped Codex TUI agents.
 .TP
 \fBloopx canary\fR
 Plan or run catalog\-informed smoke profiles.
+.TP
+\fBloopx promotion\-readiness\fR
+Record release\-scoped canary readiness evidence in the shared runtime ledger.
 .TP
 \fBloopx benchmark\fR
 Use fixture\-only benchmark runner skeletons by default.
